Moved TFWeapon out of TerrariaFortress.cs.
